Update 21415_Assignment_11.cpp
Queues are frequently used in computer programming, and a
typical example is the creation of a job queue by an operating
system. If the operating system does not use priorities, then
the jobs are processed in the order they enter the system.
Write C++ program for simulating job queue. Write
functions to add job, display job and delete job from queue..