Update 21415_Assignment_12.cpp
Write program to implement a priority queue in C++ using
an order list/array to store the items in the queue. Create a
class that includes the data items (which should be template)
and the priority (which should be int). The order list/array
should contain these objects, with operator <= overloaded so
that the items with highest priority appear at the beginning of
the list/array (which will make it relatively easy to retrieve the
highest item.)